+++ tests/test_helpers/test_bad_builtin_use.py
@@ -63,6 +63,23 @@ class TestBadBuiltinUse(dlint.test.base.BaseTest):
 
         assert result == expected
 
+    def test_bad_builtin_overwritten(self):
+        python_string = self.get_ast_node(
+            """
+            from foo import bar
+
+            result = bar()
+            """
+        )
+
+        linter = get_builtin_use_implementation('bar')
+        linter.visit(python_string)
+
+        result = linter.get_results()
+        expected = []
+
+        assert result == expected
+
     def test_no_builtin_usage(self):
         python_string = self.get_ast_node(
             """